fix: better sign out
